// frontend/src/app/services/models/user.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface User {
  address?: string;
  birth_date?: string;
  created_at?: string;
  email?: string;
  first_name?: string;
  fullName?: string;
  image?: string;
  is_deleted?: boolean;
  is_locked?: boolean;
  last_name?: string;
  password?: string;
  phone_number?: string;
  reset_pwd_code?: string;
  reset_pwd_date?: string;
  updated_at?: string;
  userId?: number;
}